but-llm: LM Studio
Add LM studio support for but-llm.

- LM studio has an OpenAI compatible API, so now both are sharing the
 client APIs and types
